Rearranging figures (#411)
* more work

* more work

* more work

* more work

* more work

* more work

* more work

* mmore work

* more work
