+++ manuscript/20.supp-figures.md
@@ -1,0 +1,45 @@
+## Supplementary Figures
+
+### Performance on uncensored lineages of increasing size
+
+[//]: # (Supp. Figure 2)
+
+![**Performance on uncensored single lineages of increasing sizes.** Single uncensored lineages with fate and time observations are simulated with two states. Lineages are created and the model's performance in estimation and state assignment accuracy are plotted. These lineages cannot represent real cells because cells that are daughters of cells that have died are observed in these plots.](./output/figureS02.svg){#fig:ideal2}
+
+[//]: # (Supp. Figure 3)
+
+![**Performance on uncensored populations of lineages of increasing number.** Populations of uncensored lineages with fate and time observations are simulated with two states. Populations (groups of lineages) are created and the model's performance in estimation and state assignment accuracy are plotted. These lineages cannot represent real cells because cells that are daughters of cells that have died are observed.](./output/figureS03.svg){#fig:ideal3}
+
+
+
+### Performance on censored lineages of increasing size
+
+[//]: # (Supp. Figure 4)
+
+![**Performance on censored single lineages of increasing sizes by increasing experiment time.** Single censored lineages with fate and time observations are simulated with two states. Cells that die or live past the desired experiment time have their subtrees (the lineages created by their possible daughters) are removed. Lineages are created and the model's performance in estimation and state assignment accuracy are plotted.](./output/figureS04.svg){#fig:censor4}
+
+[//]: # (Supp. Figure 5)
+
+![**Performance on populations of censored lineages of increasing number.** Populations of uncensored lineages with fate and time observations are simulated with two states. Cells that die or live past the desired experiment time have their subtrees (the lineages created by their possible daughters) are removed. Populations are created and the model's performance in estimation and state assignment accuracy are plotted. These lineages cannot represent real cells because cells that are daughters of cells that have died are observed.](./output/figureS05.svg){#fig:censor5}
+
+
+
+### Change in model performance when varying presence of state
+
+![**Change in model performance when varying presence of state in uncensored lineages.** Single uncensored lineages with fate and time observations are simulated with two states. The transition matrices that generate these lineages are varied such that the relative proportion of cells in each state vary. In particular, transition matrices initially bias cells to divide into only cells in the second state and force lineage trees to be entirely made of cells in the second state. A range of transition matrices are made that create lineages which bias the number of cells to be in one state or another. The proportion of cells in the first state in the lineage is recorded.](./output/figureS06.svg){#fig:prop6}
+
+![**Change in model performance when varying presence of state in censored lineages.** Single censored lineages with fate and time observations are simulated with two states. The transition matrices that generate these lineages are varied such that the relative proportion of cells in each state vary. The same experiment as before is replicated  The proportion of cells in the first state in the lineage is recorded.](./output/figureS07.svg){#fig:prop7}
+
+
+
+### Change in model performance when varying state distribution similarity
+
+![**Change in model performance when varying state distribution similarity with uncensored lineages.** Single uncensored lineages with fate and time observations are simulated with two states. These lineages are created in such a way that cells from each state are well represented in the lineage (the states are evenly distributed). One state is fixed and the other is allowed to vary in terms of its distribution. Initially, the two states are the same, i.e., they generate exactly the same distribution of cells. One of the states is then allowed to vary by increasing the gamma shape parameter. Doing so increases the Wassertein distance between the two states. Cells of each state from each lineage are sampled as distributions and are used to calculate the Wasserstein distance between the states. The Bernoulli parameter (a) and the gamma scale parameter (c) are the same for both of the states, but the gamma shape (b) parameter varies for the second state. The model is able to estimate the parameters of the states across all Wasserstein distances but has difficulty when the two states are nearly identical, when the Wasserstein distance between the states is low.](./output/figureS08.svg){#fig:wass1}
+
+![**Change in model performance when varying state distribution similarity with censored lineages.** We perform the same experiment as above but with censored lineages.](./output/figureS09.svg){#fig:wass2}
+
+
+### AIC Figure
+
+some introduction...
+![Model prediction accuracy on a simulated cell population with varying numbers of lineages. Accuracy of tHMM state assignment (left), Bernoulli parameter estimation (middle), and exponential parameter estimation (right) for populations of cells with different numbers of lineages. This is analogous to increasing the number of initial seeded cells. Lineages were composed of initially therapy sensitive (blue) that then later transitioned to resistant (orange) subpopulations. Dotted horizontal lines at target accuracy and true population values are depicted. Solid line represents a 10-point moving average. ...](./output/figureS10.svg){#fig:AIC}
